split config.rs into a config/mod.rs & schema & validation
